use a custom scope to prevent residues in global
This commit introduces a custom scope (fbr) to prevent the accumulation of residues in the global scope that can cause problems when changing a page reactively. This scope remains during the page-cycle and is dropped only after a full refresh. All classes and functions of the library are part of this scope.

In addition, this commit also adds a new update() method to the FabrCoreComponent to gain updates from the document. This method checks for changes and appends new elements to the component without re-rendering everything. This is called by Fabr for each component when a fragment changes in the document, such as updating an element in the page with remote content.

Overall, this commit improves the performance and stability of the library.